+++ idea/src/org/jetbrains/kotlin/idea/refactoring/KotlinRefactoringSupportProvider.kt
@@ -22,7 +22,9 @@ import com.intellij.psi.PsiElement
 import com.intellij.psi.PsiNameIdentifierOwner
 import com.intellij.refactoring.RefactoringActionHandler
 import org.jetbrains.kotlin.idea.refactoring.changeSignature.KotlinChangeSignatureHandler
+import org.jetbrains.kotlin.idea.refactoring.introduce.extractClass.KotlinExtractInterfaceHandler
 import org.jetbrains.kotlin.idea.refactoring.introduce.extractFunction.ExtractKotlinFunctionHandler
+import org.jetbrains.kotlin.idea.refactoring.introduce.extractClass.KotlinExtractSuperclassHandler
 import org.jetbrains.kotlin.idea.refactoring.introduce.introduceParameter.KotlinIntroduceLambdaParameterHandler
 import org.jetbrains.kotlin.idea.refactoring.introduce.introduceParameter.KotlinIntroduceParameterHandler
 import org.jetbrains.kotlin.idea.refactoring.introduce.introduceProperty.KotlinIntroducePropertyHandler
@@ -78,6 +80,10 @@ class KotlinRefactoringSupportProvider : RefactoringSupportProvider() {
     override fun getPullUpHandler() = KotlinPullUpHandler()
 
     override fun getPushDownHandler() = KotlinPushDownHandler()
+
+    override fun getExtractSuperClassHandler() = KotlinExtractSuperclassHandler
+
+    override fun getExtractInterfaceHandler() = KotlinExtractInterfaceHandler
 }
 
 class KotlinVetoRenameCondition: Condition<PsiElement> {